requirements.txt
# LlamaIndex Core y extensiones
llama-index-core
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma

# LLM (cambia si usas otro proveedor, ej. 'llama-index-llms-gemini')
llama-index-llms-openai

# Base vectorial
chromadb

# API Web
fastapi
uvicorn

# Utilidades
pydantic
python-dotenv